+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".ui.authentication.AuthActivity"
+            android:name=".ui.SplashScreenActivity"
             android:exported="true"
             android:launchMode="singleTop">
             <intent-filter>
@@ -35,9 +35,15 @@
             </intent-filter>
         </activity>
         <activity
+            android:name=".ui.authentication.AuthActivity"
+            android:exported="false"
+            android:launchMode="singleTop"
+            android:theme="@style/Theme.NuberJam"/>
+        <activity
             android:name=".ui.main.MainActivity"
             android:exported="true"
-            android:launchMode="singleTop" />
+            android:launchMode="singleTop"
+            android:theme="@style/Theme.NuberJam" />
     </application>
 
 </manifest>
