+++ app/src/main/java/com/example/nuberjam/ui/main/MainActivity.kt
@@ -6,7 +6,6 @@ import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
 import com.example.nuberjam.R
 import com.example.nuberjam.databinding.ActivityMainBinding
-import com.example.nuberjam.ui.authentication.LoginFragment
 import com.example.nuberjam.ui.customview.CustomSnackbar
 import com.google.android.material.bottomnavigation.BottomNavigationView
